requirements.txt
flask>=3.0
google-genai>=1.0
python-dotenv>=1.0
weasyprint>=62.0
